// src/app/core/models/lists.ts
import { Injectable } from '@angular/core';
export interface Listitems {
    Title: string;
    Url: string;
    Price: any;
    Category: string;
    Description: string;
    Websites: string;
    ShopAddress: string;
    LatLon: any;
    OtherInformation: string;
    OpenDate: any;
    ExpiryDate: any;
    IdUserBlock: any;
    ExpiryTimeBlock: any;
    OpenTimeBlock;
}